fix(materials): replace theme creation functions and streamline theme management
