+++ packages/materials/vue-element-plus/src/materials/materials.ts
@@ -1,7 +1,7 @@
 import { buildMaterialDefaultValueMap, type IMaterials } from '@opentiny/genui-sdk-core';
 import { materialsMeta } from '../meta';
 import { components } from './components';
-import { createElementPlusMaterialsTheme } from './theme';
+import { createTheme } from './theme';
 
 const standardRequiredCompleteFieldSelectors = ['[componentName=ElCard] > props > shadow'];
 
@@ -9,5 +9,5 @@ export const materials: IMaterials = {
   components,
   requiredCompleteFieldSelectors: standardRequiredCompleteFieldSelectors,
   defaultPropsMap: buildMaterialDefaultValueMap(materialsMeta),
-  createTheme: createElementPlusMaterialsTheme,
+  createTheme,
 };
